RFHybridController: Update to latest Spec
Updated ground station frimware to be compliant with v0.4.0 of the
Propulsion Avionics Control Plane Spec. This includes new acking
protocols, a more logical command scheme (0 is now always a noop),
formalization of packet headers/metadata, and a sequence code based
command synchonization. NACKs are pending in a future revision of the
specs.